Isolate unused code, add more comments, minor refactoring
